Move fast encoder to its own file
